# api/openapi.yaml
openapi: 3.1.0
info:
  title: App Parceiros API
  version: 0.2.0
  description: >
    Contrato da API do App Parceiros. É a fonte da verdade: altere aqui
    antes de implementar e gere os clientes a partir deste arquivo.


    Erros vêm sempre como `{code, message}` (schema `Error`): `code` é
    estável e serve para o programa; `message` já vem em português, pronta
    para mostrar ao usuário. O status segue o tipo do erro: 400 para JSON
    ou parâmetro malformado (`invalid_json`, `invalid_request`), 401 sem
    login, 402 workspace suspenso, 403 sem permissão, 404 não encontrado,
    409 conflito, 410 link vencido, 422 dado inválido, 429 limite de
    chamadas, 502 provedor externo fora do ar, 503 recurso não configurado
    no servidor e 500 `internal_error`.


    As rotas de um workspace ficam sob `/v1/workspaces/{workspaceId}`. Quem
    não é membro recebe 404 `workspace_not_found`. Com o workspace suspenso
    por falta de pagamento, só continuam valendo ver o workspace, sair dele
    (ou remover afiliados) e a assinatura; o resto responde 402
    (`workspace_suspended`, ou `workspace_suspended_owner` para o dono).
servers:
  - url: http://localhost:8080
...
paths:
...
  /v1/me:
    get:
      summary: Usuário autenticado
      description: >
        No primeiro acesso, cria o usuário a partir do perfil do provedor de
        identidade (Zitadel ou o provedor interno) e o seu workspace pessoal.
      operationId: getMe
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Usuário
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
  /v1/workspaces:
    get:
      summary: Workspaces de que o usuário participa (seletor do topo)
      operationId: listWorkspaces
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Workspaces, o pessoal primeiro
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
    post:
      summary: Criar mentoria
      description: Cria um workspace do tipo mentoria com o usuário como dono.
      operationId: createMentorship
      tags: [accounts]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateMentorship" }
      responses:
        "201":
          description: Mentoria criada
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "422":
          description: "`invalid_workspace_name` ou `invalid_photo_url`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      summary: Detalhe do workspace
      description: Continua valendo com o workspace suspenso.
      operationId: getWorkspace
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
    patch:
      summary: Alterar nome e foto (só o dono)
      operationId: updateWorkspace
      tags: [accounts]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWorkspace" }
      responses:
        "200":
          description: Workspace atualizado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "402": { $ref: "#/components/responses/WorkspaceSuspended" }
        "403":
          description: "`forbidden`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
        "422":
          description: "`invalid_workspace_name` ou `invalid_photo_url`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}/members:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      summary: Membros do workspace (dono e mentor)
      operationId: listMembers
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Membros
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Member" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "402": { $ref: "#/components/responses/WorkspaceSuspended" }
        "403":
          description: "`forbidden`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
  /v1/workspaces/{workspaceId}/members/{userId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - name: userId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Sair do workspace ou remover um membro
      description: >
        Com o próprio id, o usuário sai do workspace. Dono e mentor removem
        afiliados; só o dono remove mentores; o dono não sai nem é removido.
        Os dados pessoais do afiliado (coleção, credencial) continuam dele.
        Continua valendo com o workspace suspenso (para caber em menos
        assentos).
      operationId: removeMember
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "204": { description: Removido }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`forbidden`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "`member_not_found` ou `workspace_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`owner_cannot_leave`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}/invites:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      summary: Convites pendentes (dono e mentor)
      operationId: listInvites
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Convites ainda válidos, sem o token
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Invite" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "402": { $ref: "#/components/responses/WorkspaceSuspended" }
        "403":
          description: "`forbidden`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
    post:
      summary: Convidar afiliado (dono e mentor)
      description: >
        Gera um convite de uso único para um workspace de mentoria. Com
        e-mail, só a conta com aquele e-mail verificado pode aceitar, e o link
        também vai por e-mail (`email_sent`). Convites pendentes ocupam
        assento do plano. O token só aparece nesta resposta.
      operationId: createInvite
      tags: [accounts]
      security: [{ bearer: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateInvite" }
      responses:
        "201":
          description: Convite criado
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreatedInvite" }
        "400": { $ref: "#/components/responses/InvalidJSON" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "402": { $ref: "#/components/responses/WorkspaceSuspended" }
        "403":
          description: "`forbidden`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404": { $ref: "#/components/responses/WorkspaceNotFound" }
        "409":
          description: "`mentorship_only` ou `no_seats`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "`invalid_invite_email` ou `invalid_invite_validity`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/workspaces/{workspaceId}/invites/{inviteId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
      - name: inviteId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      summary: Cancelar convite pendente (dono e mentor)
      operationId: revokeInvite
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "204": { description: Cancelado }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "402": { $ref: "#/components/responses/WorkspaceSuspended" }
        "403":
          description: "`forbidden`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "`invite_not_found` ou `workspace_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/invites/{token}:
    parameters:
      - $ref: "#/components/parameters/InviteToken"
    get:
      summary: Ver convite (público)
      description: Mostra o workspace e a situação do convite antes do login.
      operationId: viewInvite
      tags: [accounts]
      security: []
      responses:
        "200":
          description: Convite
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PublicInvite" }
        "404":
          description: "`invite_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /v1/invites/{token}/accept:
    parameters:
      - $ref: "#/components/parameters/InviteToken"
    post:
      summary: Aceitar convite
      description: Torna o usuário afiliado do workspace do convite.
      operationId: acceptInvite
      tags: [accounts]
      security: [{ bearer: [] }]
      responses:
        "200":
          description: Workspace em que o usuário entrou
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        "401": { $ref: "#/components/responses/Unauthenticated" }
        "403":
          description: "`invite_other_email`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "404":
          description: "`invite_not_found`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "409":
          description: "`already_member` ou `no_seats`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "410":
          description: "`invite_expired`, `invite_used` ou `invite_revoked`"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }

  # --- shopee --------------------------------------------------------------
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
      description: >
        Token de acesso do provedor de identidade: JWT do Zitadel (OIDC),
        token opaco do provedor interno (`/v1/auth/login`) ou `dev:<sub>` no
        modo de desenvolvimento.
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    InviteToken:
      name: token
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthenticated:
      description: Sem token ou token inválido (`unauthenticated`)
      headers:
        WWW-Authenticate:
          schema: { type: string, const: Bearer }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    InvalidJSON:
      description: Corpo que não é JSON, com campo desconhecido ou grande demais (`invalid_json`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    WorkspaceNotFound:
      description: "`workspace_not_found`: o workspace não existe ou o usuário não é membro dele"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    WorkspaceSuspended:
      description: >
        Workspace suspenso por falta de pagamento: `workspace_suspended_owner`
        para o dono, `workspace_suspended` para os demais
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: >
            Código estável para o front. Genéricos: unauthenticated,
            internal_error, invalid_json, invalid_request. Auth (provedor
            interno): invalid_credentials, email_taken, invalid_email,
            invalid_name, weak_password, invalid_auth_token,
            too_many_attempts. Contas: workspace_not_found, member_not_found,
            invite_not_found, forbidden, mentorship_only,
            consent_mentorship_only, owner_cannot_leave, no_seats,
            seats_in_use, seats_above_plan, workspace_suspended,
            workspace_suspended_owner, already_member, invite_expired,
            invite_used, invite_revoked, invite_other_email,
            invalid_workspace_name, invalid_photo_url, invalid_invite_email,
            invalid_invite_validity. Assinatura: billing_owner_only,
            billing_managers_only, already_subscribed, no_subscription,
            invalid_tax_id, minimum_seats, no_price, simulation_unavailable,
            billing_unavailable, invalid_billing_webhook. Shopee:
            invalid_shopee_credential, shopee_access_denied,
            shopee_rate_limited, shopee_unavailable, invalid_app_id,
            invalid_secret. Catálogo e radar: product_not_found,
            invalid_page, invalid_per_page, negative_price,
            invalid_min_commission, invalid_min_rating, invalid_sort,
            query_too_long, invalid_history_days, invalid_product_link,
            short_link. Coleções: item_not_found, collection_not_found,
            collection_exists, no_credential, import_unavailable,
            link_rate_limited, product_lookup_failed,
            product_or_link_required, invalid_item_title,
            invalid_item_description, invalid_item_notes, invalid_tag,
            too_many_tags, invalid_item_status, invalid_affiliate_link,
            invalid_collection_name, too_many_item_collections,
            invalid_item_page. Notificações: notification_not_found,
            push_unavailable, invalid_push_subscription,
            too_many_push_subscriptions, invalid_notification_preferences.
            Mídia: video_not_found, invalid_video_link, video_unavailable,
            platform_unavailable, usage_rights_required,
            invalid_video_format, invalid_video_size, video_too_large,
            video_quota_exceeded, uploads_unavailable, upload_closed,
            upload_incomplete, upload_parts_required, invalid_part_number,
            video_not_ready, video_owner_only, video_share_managers_only,
            video_title_too_long, embed_not_downloadable. Curadoria:
            list_not_found, list_item_not_found, video_not_in_list,
            list_managers_only, lists_mentorship_only, list_limit,
            list_empty, already_in_list, list_full, list_not_published,
            invalid_list_title, invalid_list_description,
            invalid_list_comment, list_comment_required,
            invalid_list_order, product_not_in_list. Resultados:
            group_results_forbidden, group_results_mentorship_only,
            sync_no_credential, synced_recently, invalid_period,
            sync_credential_refused, sync_shopee_unavailable.
        message: { type: string, description: Texto em português pronto para mostrar ao usuário }
...
    User:
      type: object
      required: [id, name, email, email_verified, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        email_verified:
          type: boolean
          description: Convites por e-mail e avisos por e-mail exigem o e-mail confirmado.
        created_at: { type: string, format: date-time }
    Role:
      type: string
      enum: [owner, mentor, affiliate]
    WorkspaceKind:
      type: string
      enum: [personal, mentorship]
    Plan:
      type: string
      enum: [solo, mentorship]
    AccessStatus:
      type: string
      enum: [trial, active, free, suspended]
      description: >
        `trial` no período de avaliação, `active` com a assinatura paga,
        `free` no workspace pessoal de quem é afiliado de uma mentoria em dia
        (não é cobrado, por enquanto) e `suspended` quando o acesso venceu.
        Suspenso, o workspace só permite ver a si mesmo, sair dele (ou
        remover afiliados) e cuidar da assinatura; o resto devolve 402.
    Workspace:
      type: object
      required: [id, kind, name, photo_url, plan, status, access_until, seats, role, created_at]
      properties:
        id: { type: string, format: uuid }
        kind: { $ref: "#/components/schemas/WorkspaceKind" }
        name: { type: string }
        photo_url: { type: [string, "null"] }
        plan: { $ref: "#/components/schemas/Plan" }
        status: { $ref: "#/components/schemas/AccessStatus" }
        access_until:
          type: string
          format: date-time
          description: Até quando o workspace pode ser usado (fim do teste ou do ciclo pago, mais a tolerância).
        seats:
          type: [integer, "null"]
          description: Assentos contratados na mentoria; nulo enquanto não há pagamento.
        role: { $ref: "#/components/schemas/Role" }
        created_at: { type: string, format: date-time }
    CreateMentorship:
      type: object
      required: [name]
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        photo_url: { type: [string, "null"], format: uri, description: "URL https" }
    UpdateWorkspace:
      type: object
      additionalProperties: false
      properties:
        name: { type: string, minLength: 1, maxLength: 80 }
        photo_url:
          type: string
          description: URL https. String vazia remove a foto.
    Member:
      type: object
      required: [user_id, name, email, role, joined_at, shares_results]
      properties:
        user_id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/Role" }
        joined_at: { type: string, format: date-time }
        shares_results:
          type: boolean
          description: Se o membro autoriza ver os seus resultados agregados no painel da turma.
    CreateInvite:
      type: object
      additionalProperties: false
      properties:
        email: { type: [string, "null"], format: email }
        validity_hours:
          type: integer
          minimum: 1
          maximum: 720
          default: 168
          description: Validade do convite em horas; 0 ou ausente usa o padrão (7 dias).
    Invite:
      type: object
      required: [id, email, expires_at, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    CreatedInvite:
      allOf:
        - $ref: "#/components/schemas/Invite"
        - type: object
          required: [token, url]
          properties:
            token: { type: string, description: Só aparece na criação }
            url: { type: string, format: uri, description: "Link para enviar ao afiliado ({APP_URL}/convite/{token})" }
            email_sent:
              type: boolean
              description: Só em convites com e-mail. Falso quando o e-mail não saiu (o link continua valendo).
    InviteStatus:
      type: string
      enum: [valid, expired, used, revoked]
    PublicInvite:
      type: object
      required: [workspace_name, workspace_photo_url, status, expires_at]
      properties:
        workspace_name: { type: string }
        workspace_photo_url: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/InviteStatus" }
        expires_at: { type: string, format: date-time }

    # shopee
...